refactor(connector): Move connector mappings and endpoints to dedicated modules (#8562)
