+++ .github/CODEOWNERS
@@ -51,6 +51,9 @@ crates/api_models/src/connector_enums.rs @juspay/hyperswitch-connector
 crates/common_enums/src/connector_enums.rs @juspay/hyperswitch-connector
 crates/router/src/configs/defaults/payment_connector_required_fields.rs @juspay/hyperswitch-connector
 crates/hyperswitch_interfaces/src/configs.rs @juspay/hyperswitch-connector
+crates/hyperswitch_domain_models/src/connector_endpoints.rs @juspay/hyperswitch-connector
+crates/router/src/types/api/connector_mapping.rs @juspay/hyperswitch-connector
+crates/router/src/types/connector_transformers.rs @juspay/hyperswitch-connector
 
 crates/router/src/compatibility/ @juspay/hyperswitch-compatibility
