+++ crates/hyperswitch_domain_models/src/connector_endpoints.rs
