+++ crates/hyperswitch_domain_models/src/lib.rs
@@ -6,7 +6,7 @@ pub mod business_profile;
 pub mod callback_mapper;
 pub mod card_testing_guard_data;
 pub mod cards_info;
-pub mod configs;
+pub mod connector_endpoints;
 pub mod consts;
 pub mod customer;
 pub mod disputes;
