+++ crates/hyperswitch_domain_models/src/router_request_types/revenue_recovery.rs
@@ -1,13 +1,13 @@
 use common_enums::enums;
 
-use crate::configs;
+use crate::connector_endpoints;
 
 #[derive(Debug, Clone)]
 pub struct BillingConnectorPaymentsSyncRequest {
     /// unique id for making billing connector psync call
     pub billing_connector_psync_id: String,
     /// connector params of the connector
-    pub connector_params: configs::ConnectorParams,
+    pub connector_params: connector_endpoints::ConnectorParams,
 }
 
 #[derive(Debug, Clone)]
@@ -18,7 +18,7 @@ pub struct RevenueRecoveryRecordBackRequest {
     pub payment_method_type: Option<common_enums::PaymentMethodType>,
     pub attempt_status: common_enums::AttemptStatus,
     pub connector_transaction_id: Option<common_utils::types::ConnectorTransactionId>,
-    pub connector_params: configs::ConnectorParams,
+    pub connector_params: connector_endpoints::ConnectorParams,
 }
 
 #[derive(Debug, Clone)]
@@ -26,5 +26,5 @@ pub struct BillingConnectorInvoiceSyncRequest {
     /// Invoice id
     pub billing_connector_invoice_id: String,
     /// connector params of the connector
-    pub connector_params: configs::ConnectorParams,
+    pub connector_params: connector_endpoints::ConnectorParams,
 }
