+++ crates/hyperswitch_interfaces/src/configs.rs
@@ -1,1 +1,1 @@
-pub use hyperswitch_domain_models::configs::Connectors;
+pub use hyperswitch_domain_models::connector_endpoints::Connectors;
