+++ crates/router/src/core/revenue_recovery/types.rs
@@ -780,14 +780,15 @@ pub fn construct_recovery_record_back_router_data(
         .change_context(errors::RecoveryError::RecordBackToBillingConnectorFailed)
         .attach_printable("Cannot find connector from the connector_name")?;
 
-    let connector_params = hyperswitch_domain_models::configs::Connectors::get_connector_params(
-        &state.conf.connectors,
-        connector,
-    )
-    .change_context(errors::RecoveryError::RecordBackToBillingConnectorFailed)
-    .attach_printable(format!(
-        "cannot find connector params for this connector {connector} in this flow",
-    ))?;
+    let connector_params =
+        hyperswitch_domain_models::connector_endpoints::Connectors::get_connector_params(
+            &state.conf.connectors,
+            connector,
+        )
+        .change_context(errors::RecoveryError::RecordBackToBillingConnectorFailed)
+        .attach_printable(format!(
+            "cannot find connector params for this connector {connector} in this flow",
+        ))?;
 
     let router_data = router_data_v2::RouterDataV2 {
         flow: PhantomData::<router_flow_types::RecoveryRecordBack>,
