+++ crates/router/src/core/webhooks/recovery_incoming.rs
@@ -977,7 +977,7 @@ impl BillingConnectorPaymentsSyncFlowRouterData {
             .attach_printable("Cannot find connector from the connector_name")?;
 
         let connector_params =
-            hyperswitch_domain_models::configs::Connectors::get_connector_params(
+            hyperswitch_domain_models::connector_endpoints::Connectors::get_connector_params(
                 &state.conf.connectors,
                 connector,
             )
@@ -1143,7 +1143,7 @@ impl BillingConnectorInvoiceSyncFlowRouterData {
             .attach_printable("Cannot find connector from the connector_name")?;
 
         let connector_params =
-            hyperswitch_domain_models::configs::Connectors::get_connector_params(
+            hyperswitch_domain_models::connector_endpoints::Connectors::get_connector_params(
                 &state.conf.connectors,
                 connector,
             )
